requirements.txt
fastmcp==2.12.4
PyYAML==6.0.3
httpx==0.28.1
